+++ apps/renderer/src/modules/customize-toolbar/constant.ts
@@ -1,0 +1,36 @@
+import type { UniqueIdentifier } from "@dnd-kit/core"
+
+import { COMMAND_ID } from "../command/commands/id"
+
+export interface ToolbarActionOrder {
+  main: UniqueIdentifier[]
+  more: UniqueIdentifier[]
+}
+
+export const DEFAULT_ACTION_ORDER: ToolbarActionOrder = {
+  main: Object.values(COMMAND_ID.entry).filter(
+    (id) =>
+      !(
+        [
+          COMMAND_ID.entry.read,
+          COMMAND_ID.entry.unread,
+          COMMAND_ID.entry.copyLink,
+          COMMAND_ID.entry.openInBrowser,
+        ] as string[]
+      ).includes(id),
+  ),
+  more: [
+    ...Object.values(COMMAND_ID.integration),
+    ...Object.values(COMMAND_ID.entry).filter((id) =>
+      (
+        [
+          COMMAND_ID.entry.read,
+          COMMAND_ID.entry.unread,
+          COMMAND_ID.entry.copyLink,
+          COMMAND_ID.entry.openInBrowser,
+        ] as string[]
+      ).includes(id),
+    ),
+    COMMAND_ID.settings.customizeToolbar,
+  ],
+}
