+++ packages/shared/src/interface/settings.ts
@@ -52,6 +52,12 @@ export interface UISettings {
   pictureViewMasonry: boolean
   pictureViewFilterNoImage: boolean
   wideMode: boolean
+
+  // Action Order
+  toolbarOrder: {
+    main: (string | number)[]
+    more: (string | number)[]
+  }
 }
 
 export interface IntegrationSettings {
